# api/openapi.json
{
  "components": {
    "schemas": {
      "AiDatasetColumnIn": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 255,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "maxLength": 128,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "AiDatasetColumnIn",
        "type": "object"
      },
      "AiDatasetIn": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/AiDatasetColumnIn"
            },
            "maxItems": 500,
            "title": "Columns",
            "type": "array"
          },
          "table": {
            "maxLength": 255,
            "title": "Table",
            "type": "string"
          }
        },
        "required": [
          "table",
          "columns"
        ],
        "title": "AiDatasetIn",
        "type": "object"
      },
...
      "ConverseIn": {
        "properties": {
          "focusTable": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focustable"
          },
          "kind": {
            "enum": [
              "query",
              "dashboard",
              "kpi",
              "report",
              "snippet"
            ],
            "title": "Kind",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ConverseMessageIn"
            },
            "maxItems": 25,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "targetDashboardId": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetdashboardid"
          }
        },
        "required": [
          "kind",
          "messages"
        ],
        "title": "ConverseIn",
        "type": "object"
      },
      "ConverseMessageIn": {
        "properties": {
          "content": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ConverseMessageIn",
        "type": "object"
      },
      "ConverseOut": {
        "properties": {
          "focusTable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focustable"
          },
          "proposal": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "dashboard": "#/components/schemas/DashboardProposalOut",
                    "kpi": "#/components/schemas/KpiProposalOut",
                    "query": "#/components/schemas/QueryProposalOut",
                    "report": "#/components/schemas/ReportProposalOut",
                    "snippet": "#/components/schemas/SnippetProposalOut"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/QueryProposalOut"
                  },
                  {
                    "$ref": "#/components/schemas/DashboardProposalOut"
                  },
                  {
                    "$ref": "#/components/schemas/KpiProposalOut"
                  },
                  {
                    "$ref": "#/components/schemas/ReportProposalOut"
                  },
                  {
                    "$ref": "#/components/schemas/SnippetProposalOut"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "suggestedAnswers": {
            "items": {
              "type": "string"
            },
            "maxItems": 4,
            "title": "Suggestedanswers",
            "type": "array"
          }
        },
        "required": [
          "reply",
          "suggestedAnswers",
          "ready",
          "proposal",
          "focusTable"
        ],
        "title": "ConverseOut",
        "type": "object"
      },
      "DashboardProposalOut": {
        "properties": {
          "kind": {
            "const": "dashboard",
            "default": "dashboard",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "widgets": {
            "items": {
              "$ref": "#/components/schemas/DashboardWidgetProposalOut"
            },
            "title": "Widgets",
            "type": "array"
          }
        },
        "required": [
          "name",
          "widgets"
        ],
        "title": "DashboardProposalOut",
        "type": "object"
      },
...
      "DomainHubOut": {
        "properties": {
          "counters": {
            "items": {
              "$ref": "#/components/schemas/HubCounterOut"
            },
            "title": "Counters",
            "type": "array"
          },
          "dashboardIds": {
            "items": {
              "type": "integer"
            },
            "title": "Dashboardids",
            "type": "array"
          },
          "datasetIds": {
            "items": {
              "type": "string"
            },
            "title": "Datasetids",
            "type": "array"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "datasetIds",
          "dashboardIds",
          "counters"
        ],
        "title": "DomainHubOut",
        "type": "object"
      },
...
      "GenerateSqlIn": {
        "properties": {
          "currentSql": {
            "anyOf": [
              {
                "maxLength": 50000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentsql"
          },
          "dataset": {
            "$ref": "#/components/schemas/AiDatasetIn"
          },
          "prompt": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt",
          "dataset"
        ],
        "title": "GenerateSqlIn",
        "type": "object"
      },
      "GenerateSqlOut": {
        "properties": {
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "sql": {
            "title": "Sql",
            "type": "string"
          }
        },
        "required": [
          "sql",
          "rationale"
        ],
        "title": "GenerateSqlOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HubCounterOut": {
        "properties": {
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "deltaUnit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltaunit"
          },
          "kpiId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kpiid"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "queryId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queryid"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "title": "HubCounterOut",
        "type": "object"
      },
      "KpiProposalOut": {
        "description": "A KPI over a query that exists, or over one written for it.\n\nA KPI reads one number from one query, so it needs exactly one source. Both\nhalves are always on the wire with the unused one null, the same rule\nDashboardWidgetProposalOut follows and for the same reason.",
        "properties": {
          "atRisk": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atrisk"
          },
          "breached": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breached"
          },
          "cadence": {
            "enum": [
              "hourly",
              "daily",
              "weekly"
            ],
            "title": "Cadence",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "direction": {
            "enum": [
              "higher-is-better",
              "lower-is-better"
            ],
            "title": "Direction",
            "type": "string"
          },
          "kind": {
            "const": "kpi",
            "default": "kpi",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "newQuery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewQueryProposalOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "sourceQueryId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcequeryid"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "valueColumn": {
            "title": "Valuecolumn",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "sourceQueryId",
          "newQuery",
          "valueColumn",
          "unit",
          "target",
          "direction",
          "cadence",
          "atRisk",
          "breached"
        ],
        "title": "KpiProposalOut",
        "type": "object"
      },
...
      "QueryProposalOut": {
        "properties": {
          "datasetTable": {
            "title": "Datasettable",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "kind": {
            "const": "query",
            "default": "query",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sql": {
            "title": "Sql",
            "type": "string"
          },
          "vizChoiceId": {
            "title": "Vizchoiceid",
            "type": "string"
          },
          "vizOptions": {
            "additionalProperties": true,
            "title": "Vizoptions",
            "type": "object"
          }
        },
        "required": [
          "name",
          "description",
          "sql",
          "datasetTable",
          "vizChoiceId",
          "vizOptions"
        ],
        "title": "QueryProposalOut",
        "type": "object"
      },
      "ReportOutlineOut": {
        "properties": {
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/OutlineSectionOut"
            },
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "goal",
          "sections"
        ],
        "title": "ReportOutlineOut",
        "type": "object"
      },
      "ReportProposalOut": {
        "properties": {
          "kind": {
            "const": "report",
            "default": "report",
            "title": "Kind",
            "type": "string"
          },
          "outline": {
            "$ref": "#/components/schemas/ReportOutlineOut"
          }
        },
        "required": [
          "outline"
        ],
        "title": "ReportProposalOut",
        "type": "object"
      },
      "SnippetProposalOut": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "kind": {
            "const": "snippet",
            "default": "snippet",
            "title": "Kind",
            "type": "string"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "trigger",
          "snippet",
          "description"
        ],
        "title": "SnippetProposalOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "veodyn-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/ai/converse": {
      "post": {
        "description": "One Create-with-AI interview turn, grounded on what actually exists.\n\nThe request carries no grounding, unlike /outline: the catalog and the query\nlist are assembled here, so a modified client cannot make the model believe\nin a table that does not exist.",
        "operationId": "post_converse_ai_converse_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConverseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConverseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Converse",
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/generate-sql": {
      "post": {
        "description": "Natural language to one read-only SELECT over the given dataset.\n\nGiven the same warehouse profile the interview gets, so the statement is\nwritten against real cardinality and real ranges.",
        "operationId": "post_generate_sql_ai_generate_sql_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSqlIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateSqlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Generate Sql",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/domains": {
      "get": {
        "operationId": "list_domain_hubs_domains_get",
        "parameters": [
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DomainHubOut"
                  },
                  "title": "Response List Domain Hubs Domains Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Domain Hubs",
        "tags": [
          "catalog"
        ]
      }
    },
    "/domains/{key}": {
      "get": {
        "operationId": "get_domain_hub_domains__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainHubOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Domain Hub",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}